+++ resources/compiledJS/arrays/items.js
@@ -306,7 +306,7 @@ const items = {
         price: 800,
         weight: 1.7,
         type: "weapon",
-        firesProjectile: "iceSpikeProjectile",
+        firesProjectile: "iceSpikedProjectile",
         statBonus: "int",
         grade: "rare",
         slot: "weapon",