+++ script/localisation.ts
@@ -259,7 +259,7 @@ const finnish = {
   rangerArmor_name: "Metsästäjän Haarniska",
   rangerBoots_name: "Metsästäjän Kengät",
   rangerPants_name: "Metsästäjän Housut",
-  rangerGloves_name: "Metästäjän Hanskat",
+  rangerGloves_name: "Metsästäjän Hanskat",
   rangerHood_name: "Metsästäjän Huppu",
   knightArmor_name: "Ritarin Haarniska",
   knightSabatons_name: "Ritarin Jalkimet",
@@ -368,6 +368,7 @@ const finnish = {
   // Dialog screen stuff
   openStore: "KAUPPA",
   quest: "TEHTÄVÄ",
+  smith: "SEPPÄ",
   exit: "POISTU",
 
   // Misc
@@ -444,6 +445,7 @@ const finnish = {
 
   // Character names
   testMerchant_name: "Karl Bauer",
+  blacksmithMaroch_name: "Maroch",
 
   // Passive abilities
   resilience_of_the_lone_wanderer_name: "Kulkijan Sinnikkyys",
@@ -829,7 +831,7 @@ const finnish = {
   smoke_screen_desc: "Peitä vastustajasi näkökenttä.",
   sneakier_stabbing_name: "Konnan Kepponen",
   sneakier_stabbing_desc: "Tästä tempusta eivät puukotukselle allergiset pidä!",
-  smoke_and_mirrors_name: "Taika Temppu",
+  smoke_and_mirrors_name: "Taikatemppu",
   smoke_and_mirrors_desc: "Savua ja peilejä...",
   tricky_distraction_name: "Harhautusliike",
   tricky_distraction_desc: "Laita vastustajasi läiskimään väärää kohdetta.",
@@ -1175,6 +1177,7 @@ const english = {
   // Dialog screen stuff
   openStore: "STORE",
   quest: "QUEST",
+  smith: "SMITH",
   exit: "EXIT",
 
   // Misc
@@ -1251,6 +1254,7 @@ const english = {
 
   // Character names
   testMerchant_name: "Karl Bauer",
+  blacksmithMaroch_name: "Maroch",
 
   // Artifacts
   part_of_set: "Part of",